Basic swap works, need to update tests that add kernels manually
