+++ modules/solid_mechanics/src/kernels/lagrangian/HomogenizedTotalLagrangianStressDivergence.C
@@ -9,6 +9,10 @@
 
 #include "HomogenizedTotalLagrangianStressDivergence.h"
 
+// MOOSE includes
+#include "Function.h"
+#include "MooseVariableScalar.h"
+
 registerMooseObject("SolidMechanicsApp", HomogenizedTotalLagrangianStressDivergence);
 
 InputParameters
@@ -17,64 +21,242 @@ HomogenizedTotalLagrangianStressDivergence::validParams()
   InputParameters params = TotalLagrangianStressDivergence::validParams();
   params.addClassDescription("Total Lagrangian stress equilibrium kernel with "
                              "homogenization constraint Jacobian terms");
-  params.addRequiredCoupledVar("macro_gradient", "Optional scalar field with the macro gradient");
-  params.addRequiredParam<UserObjectName>("homogenization_constraint",
-                                          "The UserObject defining the homogenization constraint");
+  params.renameCoupledVar(
+      "scalar_variable", "macro_var", "Optional scalar field with the macro gradient");
+  params.addRequiredParam<MultiMooseEnum>(
+      "constraint_types",
+      Homogenization::constraintType,
+      "Type of each constraint: strain, stress, or none. The types are specified in the "
+      "column-major order, and there must be 9 entries in total.");
+  params.addRequiredParam<std::vector<FunctionName>>(
+      "targets", "Functions giving the targets to hit for constraint types that are not none.");
 
   return params;
 }
 
 HomogenizedTotalLagrangianStressDivergence::HomogenizedTotalLagrangianStressDivergence(
     const InputParameters & parameters)
-  : TotalLagrangianStressDivergence(parameters),
-    _macro_gradient_num(coupledScalar("macro_gradient")),
-    _constraint(getUserObject<HomogenizationConstraint>("homogenization_constraint")),
-    _cmap(_constraint.getConstraintMap())
+  : TotalLagrangianStressDivergence(parameters)
 {
+  // Constraint types
+  auto types = getParam<MultiMooseEnum>("constraint_types");
+  if (types.size() != Moose::dim * Moose::dim)
+    mooseError("Number of constraint types must equal dim * dim. ", types.size(), " are provided.");
+
+  // Targets to hit
+  const std::vector<FunctionName> & fnames = getParam<std::vector<FunctionName>>("targets");
+
+  // Prepare the constraint map
+  unsigned int fcount = 0;
+  for (const auto j : make_range(Moose::dim))
+    for (const auto i : make_range(Moose::dim))
+    {
+      const auto idx = i + Moose::dim * j;
+      const auto ctype = static_cast<Homogenization::ConstraintType>(types.get(idx));
+      if (ctype != Homogenization::ConstraintType::None)
+      {
+        const Function * const f = &getFunctionByName(fnames[fcount++]);
+        _cmap[{i, j}] = {ctype, f};
+      }
+    }
+}
+
+std::set<std::string>
+HomogenizedTotalLagrangianStressDivergence::additionalROVariables()
+{
+  // Add the scalar variable to the list of variables that this kernel contributes to
+  std::set<std::string> vars = TotalLagrangianStressDivergence::additionalROVariables();
+  vars.insert(_kappa_var_ptr->name());
+  return vars;
 }
 
 void
-HomogenizedTotalLagrangianStressDivergence::computeOffDiagJacobianScalar(unsigned int jvar)
+HomogenizedTotalLagrangianStressDivergence::computeScalarResidual()
 {
-  if (jvar == _macro_gradient_num)
-  {
-    prepareMatrixTag(_assembly, _var.number(), jvar, _ken);
-    prepareMatrixTag(_assembly, jvar, _var.number(), _kne);
+  std::vector<Real> scalar_residuals(_k_order);
 
+  // only assemble scalar residual once; i.e. when handling the first displacement component
+  if (_alpha == 0)
+  {
     for (_qp = 0; _qp < _qrule->n_points(); _qp++)
     {
-      const Real dV = _JxW[_qp] * _coord[_qp];
-      unsigned int h = 0;
+      initScalarQpResidual();
+      Real dV = _JxW[_qp] * _coord[_qp];
+      _h = 0; // single index for residual vector; double indices for constraint tensor component
       for (auto && [indices, constraint] : _cmap)
       {
         auto && [i, j] = indices;
-        auto && ctype = constraint.first;
-        for (_i = 0; _i < _test.size(); _i++)
+        auto && [ctype, ctarget] = constraint;
+
+        if (_large_kinematics)
         {
-          // This assumes Galerkin, i.e. the test and trial functions are the same
-          _j = _i;
-
-          // Base Jacobian
-          _ken(_i, h) += _dpk1[_qp].contractionKl(i, j, gradTest(_alpha)) * dV;
-
-          // Constraint Jacobian
           if (ctype == Homogenization::ConstraintType::Stress)
-            _kne(h, _i) += _dpk1[_qp].contractionIj(i, j, gradTrial(_alpha)) * dV;
+            scalar_residuals[_h++] += dV * (_pk1[_qp](i, j) - ctarget->value(_t, _q_point[_qp]));
           else if (ctype == Homogenization::ConstraintType::Strain)
-            if (_large_kinematics)
-              _kne(h, _i) += Real(i == _alpha) * gradTrial(_alpha)(i, j) * dV;
-            else
-              _kne(h, _i) += 0.5 *
-                             (Real(i == _alpha) * gradTrial(_alpha)(i, j) +
-                              Real(j == _alpha) * gradTrial(_alpha)(j, i)) *
-                             dV;
+            scalar_residuals[_h++] +=
+                dV * (_F[_qp](i, j) - (Real(i == j) + ctarget->value(_t, _q_point[_qp])));
           else
-            mooseError("Unknown constraint type in kernel calculation!");
+            mooseError("Unknown constraint type in the integral!");
         }
-        h++;
+        else
+        {
+          if (ctype == Homogenization::ConstraintType::Stress)
+            scalar_residuals[_h++] += dV * (_pk1[_qp](i, j) - ctarget->value(_t, _q_point[_qp]));
+          else if (ctype == Homogenization::ConstraintType::Strain)
+            scalar_residuals[_h++] += dV * (0.5 * (_F[_qp](i, j) + _F[_qp](j, i)) -
+                                            (Real(i == j) + ctarget->value(_t, _q_point[_qp])));
+          else
+            mooseError("Unknown constraint type in the integral!");
+        }
       }
     }
-    accumulateTaggedLocalMatrix(_assembly, _var.number(), jvar, _ken);
-    accumulateTaggedLocalMatrix(_assembly, jvar, _var.number(), _kne);
   }
+
+  addResiduals(
+      _assembly, scalar_residuals, _kappa_var_ptr->dofIndices(), _kappa_var_ptr->scalingFactor());
+}
+
+void
+HomogenizedTotalLagrangianStressDivergence::computeScalarJacobian()
+{
+  _local_ke.resize(_k_order, _k_order);
+
+  // only assemble scalar residual once; i.e. when handling the first displacement component
+  if (_alpha == 0)
+  {
+    for (_qp = 0; _qp < _qrule->n_points(); _qp++)
+    {
+      initScalarQpJacobian(_kappa_var);
+      Real dV = _JxW[_qp] * _coord[_qp];
+
+      _h = 0;
+      for (auto && [indices1, constraint1] : _cmap)
+      {
+        auto && [i, j] = indices1;
+        auto && ctype = constraint1.first;
+        _l = 0;
+        for (auto && indices2 : _cmap)
+        {
+          auto && [a, b] = indices2.first;
+          if (ctype == Homogenization::ConstraintType::Stress)
+            _local_ke(_h, _l++) += dV * (_dpk1[_qp](i, j, a, b));
+          else if (ctype == Homogenization::ConstraintType::Strain)
+          {
+            if (_large_kinematics)
+              _local_ke(_h, _l++) += dV * (Real(i == a && j == b));
+            else
+              _local_ke(_h, _l++) +=
+                  dV * (0.5 * Real(i == a && j == b) + 0.5 * Real(i == b && j == a));
+          }
+          else
+            mooseError("Unknown constraint type in Jacobian calculator!");
+        }
+        _h++;
+      }
+    }
+  }
+
+  addJacobian(_assembly,
+              _local_ke,
+              _kappa_var_ptr->dofIndices(),
+              _kappa_var_ptr->dofIndices(),
+              _kappa_var_ptr->scalingFactor());
+}
+
+void
+HomogenizedTotalLagrangianStressDivergence::computeScalarOffDiagJacobian(
+    const unsigned int jvar_num)
+{
+  const auto & jvar = getVariable(jvar_num);
+  // Get dofs and order of this variable; at least one will be _var
+  const auto jvar_size = jvar.phiSize();
+  _local_ke.resize(_k_order, jvar_size);
+
+  for (_qp = 0; _qp < _qrule->n_points(); _qp++)
+  {
+    // single index for Jacobian column; double indices for constraint tensor component
+    unsigned int h = 0;
+    Real dV = _JxW[_qp] * _coord[_qp];
+    for (auto && [indices, constraint] : _cmap)
+    {
+      // copy constraint indices to protected variables to pass to Qp routine
+      _m = indices.first;
+      _n = indices.second;
+      _ctype = constraint.first;
+      initScalarQpOffDiagJacobian(jvar);
+      for (_j = 0; _j < jvar_size; _j++)
+        _local_ke(h, _j) += dV * computeScalarQpOffDiagJacobian(jvar_num);
+      h++;
+    }
+  }
+
+  addJacobian(_assembly,
+              _local_ke,
+              _kappa_var_ptr->dofIndices(),
+              jvar.dofIndices(),
+              _kappa_var_ptr->scalingFactor());
+}
+
+void
+HomogenizedTotalLagrangianStressDivergence::computeOffDiagJacobianScalarLocal(
+    const unsigned int svar_num)
+{
+
+  // Get dofs and order of this scalar; at least one will be _kappa_var
+  const auto & svar = _sys.getScalarVariable(_tid, svar_num);
+  const unsigned int s_order = svar.order();
+  _local_ke.resize(_test.size(), s_order);
+
+  for (_qp = 0; _qp < _qrule->n_points(); _qp++)
+  {
+    unsigned int l = 0;
+    Real dV = _JxW[_qp] * _coord[_qp];
+    for (auto && [indices, constraint] : _cmap)
+    {
+      // copy constraint indices to protected variables to pass to Qp routine
+      _m = indices.first;
+      _n = indices.second;
+      _ctype = constraint.first;
+      initScalarQpJacobian(svar_num);
+      for (_i = 0; _i < _test.size(); _i++)
+      {
+        _local_ke(_i, l) += dV * computeQpOffDiagJacobianScalar(svar_num);
+      }
+      l++;
+    }
+  }
+
+  addJacobian(_assembly, _local_ke, _var.dofIndices(), svar.dofIndices(), _var.scalingFactor());
+}
+
+Real
+HomogenizedTotalLagrangianStressDivergence::computeQpOffDiagJacobianScalar(unsigned int svar_num)
+{
+  // Just in case, skip any other scalar variables
+  if (svar_num == _kappa_var)
+    return _dpk1[_qp].contractionKl(_m, _n, gradTest(_alpha));
+  else
+    return 0.;
+}
+
+Real
+HomogenizedTotalLagrangianStressDivergence::computeScalarQpOffDiagJacobian(unsigned int jvar_num)
+{
+  // ONLY assemble the contribution from _alpha component, which is connected with _var
+  // The other components are handled by other kernel instances with other _alpha
+  if (jvar_num == _var.number())
+  {
+    if (_ctype == Homogenization::ConstraintType::Stress)
+      return _dpk1[_qp].contractionIj(_m, _n, gradTrial(_alpha));
+    else if (_ctype == Homogenization::ConstraintType::Strain)
+      if (_large_kinematics)
+        return Real(_m == _alpha) * gradTrial(_alpha)(_m, _n);
+      else
+        return 0.5 * (Real(_m == _alpha) * gradTrial(_alpha)(_m, _n) +
+                      Real(_n == _alpha) * gradTrial(_alpha)(_n, _m));
+    else
+      mooseError("Unknown constraint type in kernel calculation!");
+  }
+  else
+    return 0.;
 }
