--- modules/solid_mechanics/test/tests/lagrangian/cartesian/total/homogenization/scalar_kernel/gold/2d-mixed_out.csv
@@ -1,7 +1,0 @@
-time,exx,exy,eyy,sxx,sxy,syy,hvar_0,hvar_1,hvar_2
-0,0,0,0,0,0,0,0,0,0
-0.2,0.0027054482078334,0.00030389022096814,-0.004,79.999999999998,20,-370.94221465096,0.0027054482078335,0.00060778044193629,-0.004
-0.4,0.0054108964156669,0.00060778044193629,-0.008,160,40,-741.88442930191,0.0054108964156669,0.0012155608838726,-0.008
-0.6,0.0081163446235004,0.00091167066290443,-0.012,240.00000000001,60,-1112.8266439529,0.0081163446235003,0.0018233413258089,-0.012
-0.8,0.010821792831334,0.0012155608838726,-0.016,320,80,-1483.7688586038,0.010821792831334,0.0024311217677452,-0.016
-1,0.013527241039167,0.0015194511048407,-0.02,399.99999999998,100,-1854.7110732548,0.013527241039167,0.0030389022096814,-0.02
