--- modules/solid_mechanics/test/tests/lagrangian/cartesian/total/homogenization/scalar_kernel/gold/2d-strain_out.csv
@@ -1,7 +1,0 @@
-time,exx,exy,eyy,sxx,sxy,syy,hvar_0,hvar_1,hvar_2
-0,0,0,0,0,0,0,0,0,0
-0.2,0.008,0.002,-0.004,896.5686634732,131.57865055519,75.455538893295,0.008,0.004,-0.004
-0.4,0.016,0.004,-0.008,1793.1373269464,263.15730111037,150.91107778658,0.016,0.008,-0.008
-0.6,0.024,0.006,-0.012,2689.7059904196,394.73595166556,226.36661667988,0.024,0.012,-0.012
-0.8,0.032,0.008,-0.016,3586.2746538928,526.31460222074,301.82215557317,0.032,0.016,-0.016
-1,0.04,0.01,-0.02,4482.843317366,657.89325277593,377.27769446646,0.04,0.02,-0.02
