--- modules/solid_mechanics/test/tests/lagrangian/cartesian/total/homogenization/scalar_kernel/gold/2drow_out.csv
@@ -1,7 +1,0 @@
-time,exx,exy,eyy,sxx,sxy,syy,hvar,hvarA_0,hvarA_1
-0,0,0,0,0,0,0,0,0,0
-0.2,0.00096026311741124,0.00030401429687746,-0.00080769950780423,80.000000000007,20,-39.999999999991,0.00096026311741122,0.00060802859375492,-0.00080769950780428
-0.4,0.0019205262348225,0.00060802859375492,-0.0016153990156086,160.00000000001,40,-80,0.0019205262348224,0.0012160571875098,-0.0016153990156086
-0.6,0.0028807893522336,0.00091204289063238,-0.0024230985234128,239.99999999999,60,-120.00000000001,0.0028807893522336,0.0018240857812648,-0.0024230985234128
-0.8,0.0038410524696449,0.0012160571875098,-0.0032307980312172,319.99999999999,80,-160.00000000001,0.0038410524696449,0.0024321143750197,-0.0032307980312171
-1,0.0048013155870561,0.0015200714843873,-0.0040384975390213,400,100,-199.99999999999,0.0048013155870561,0.0030401429687746,-0.0040384975390213
